Splitting operator assertions outside StreamAssert to MessageStreamAssert, addressing review, renaming utils
