+++ samza-test/src/test/java/org/apache/samza/test/framework/StreamTaskIntegrationTest.java
@@ -78,13 +78,13 @@ public class StreamTaskIntegrationTest {
         .addOverrideConfig("job.container.thread.pool.size", "4")
         .run(Duration.ofSeconds(1));
 
-    StreamAssert.that(output).contains(outputList, Duration.ofMillis(1000));
+    StreamAssert.containsInOrder(output, outputList, Duration.ofMillis(1000));
   }
 
   @Test
   public void testSyncTaskWithMultiplePartition() throws Exception {
-    Map<Integer, List<KV>> input = new HashMap<>();
-    Map<Integer, List<Integer>> output = new HashMap<>();
+    Map<Integer, List<KV>> inputPartitionData = new HashMap<>();
+    Map<Integer, List<Integer>> expectedOutputPartitionData = new HashMap<>();
     List<Integer> partition = Arrays.asList(1, 2, 3, 4, 5);
     List<Integer> outputPartition = partition.stream().map(x -> x * 10).collect(Collectors.toList());
     for (int i = 0; i < 5; i++) {
@@ -92,11 +92,11 @@ public class StreamTaskIntegrationTest {
       for (Integer val : partition) {
         keyedPartition.add(KV.of(i, val));
       }
-      input.put(i, keyedPartition);
-      output.put(i, new ArrayList<Integer>(outputPartition));
+      inputPartitionData.put(i, keyedPartition);
+      expectedOutputPartitionData.put(i, new ArrayList<Integer>(outputPartition));
     }
 
-    CollectionStream<KV> inputStream = CollectionStream.of("test", "input", input);
+    CollectionStream<KV> inputStream = CollectionStream.of("test", "input", inputPartitionData);
     CollectionStream outputStream = CollectionStream.empty("test", "output", 5);
 
     TestRunner
@@ -105,13 +105,13 @@ public class StreamTaskIntegrationTest {
         .addOutputStream(outputStream)
         .run(Duration.ofSeconds(2));
 
-    StreamAssert.that(outputStream).contains(output, Duration.ofMillis(1000));
+    StreamAssert.containsInOrder(outputStream, expectedOutputPartitionData, Duration.ofMillis(1000));
   }
 
   @Test
   public void testSyncTaskWithMultiplePartitionMultithreaded() throws Exception {
-    Map<Integer, List<KV>> input = new HashMap<>();
-    Map<Integer, List<Integer>> output = new HashMap<>();
+    Map<Integer, List<KV>> inputPartitionData = new HashMap<>();
+    Map<Integer, List<Integer>> expectedOutputPartitionData = new HashMap<>();
     List<Integer> partition = Arrays.asList(1, 2, 3, 4, 5);
     List<Integer> outputPartition = partition.stream().map(x -> x * 10).collect(Collectors.toList());
     for (int i = 0; i < 5; i++) {
@@ -119,11 +119,11 @@ public class StreamTaskIntegrationTest {
       for (Integer val : partition) {
         keyedPartition.add(KV.of(i, val));
       }
-      input.put(i, keyedPartition);
-      output.put(i, new ArrayList<Integer>(outputPartition));
+      inputPartitionData.put(i, keyedPartition);
+      expectedOutputPartitionData.put(i, new ArrayList<Integer>(outputPartition));
     }
 
-    CollectionStream<KV> inputStream = CollectionStream.of("test", "input", input);
+    CollectionStream<KV> inputStream = CollectionStream.of("test", "input", inputPartitionData);
     CollectionStream outputStream = CollectionStream.empty("test", "output", 5);
 
     TestRunner
@@ -133,6 +133,6 @@ public class StreamTaskIntegrationTest {
         .addOverrideConfig("job.container.thread.pool.size", "4")
         .run(Duration.ofSeconds(2));
 
-    StreamAssert.that(outputStream).contains(output, Duration.ofMillis(1000));
+    StreamAssert.containsInOrder(outputStream, expectedOutputPartitionData, Duration.ofMillis(1000));
   }
 }
