+++ samza-test/src/test/java/org/apache/samza/test/operator/StreamApplicationIntegrationTestHarness.java
@@ -34,8 +34,8 @@ import org.apache.samza.execution.TestStreamManager;
 import org.apache.samza.runtime.AbstractApplicationRunner;
 import org.apache.samza.runtime.ApplicationRunner;
 import org.apache.samza.system.kafka.KafkaSystemAdmin;
+import org.apache.samza.test.framework.MessageStreamAssert;
 import org.apache.samza.test.harness.AbstractIntegrationTestHarness;
-import org.apache.samza.test.framework.StreamAssert;
 import scala.Option;
 import scala.Option$;
 
@@ -260,7 +260,7 @@ public class StreamApplicationIntegrationTestHarness extends AbstractIntegration
     AbstractApplicationRunner runner = (AbstractApplicationRunner) ApplicationRunner.fromConfig(config);
     runner.run(streamApplication);
 
-    StreamAssert.waitForComplete();
+    MessageStreamAssert.waitForComplete();
     return new RunApplicationContext(runner, config);
   }
